Relocate: Moved scrollToTop to src/components/shared as requested
